% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpf_filter.R
\name{bpf_filter}
\alias{bpf_filter}
\title{bpf_filter}
\usage{
bpf_filter(downsample_data = c(), verbose = FALSE)
}
\arguments{
\item{downsample_data}{Matrix containing downsampled data (as returned from agcounts::resample)}

\item{verbose}{Print diagnostic messages}
}
\value{
The filtered data
}
\description{
Bandpass filter for actigraph downsampled data
}
