% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{factors}
\alias{factors}
\title{factors}
\usage{
factors(sf)
}
\arguments{
\item{sf}{Integer within seq(30, 100, by = 10)}
}
\value{
upsample and downsample factors
}
\description{
Given a sampling frequency, returns the upsample and downsample factors to
set the signal to 30 Hz
}
\examples{
# not run
factors(60)
# upsample downsample
# 1          2
}
