% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts}
\alias{get_counts}
\title{get_counts}
\usage{
get_counts(raw, sf, epoch, lfe_select = FALSE, verbose = FALSE)
}
\arguments{
\item{raw}{Matrix containing raw data}

\item{sf}{Sample frequency of raw data (Hz)}

\item{epoch}{Epoch length to aggregate activity counts}

\item{lfe_select}{False for regular trimming, True for allow more noise}

\item{verbose}{Print diagnostic messages}
}
\value{
Matrix containing the count values per epoch
}
\description{
get_counts
}
\references{
{https://www.researchsquare.com/article/rs-1370418/v1}
}
\author{
Jairo Hidalgo Migueles
}
