% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{resample}
\usage{
resample(raw = c(), sf = 30, verbose = FALSE)
}
\arguments{
\item{raw}{Matrix containing raw data}

\item{sf}{Sample frequency of raw data (Hz)}

\item{verbose}{Print diagnostic messages}
}
\value{
resampled_data
}
\description{
Resample the raw data.
}
