% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_10hz.R
\name{resample_10hz}
\alias{resample_10hz}
\title{resample_10hz}
\usage{
resample_10hz(trim_data = c(), verbose = FALSE)
}
\arguments{
\item{trim_data}{Matrix containing the trimmed/thresholded data}

\item{verbose}{Print diagnostic messages}
}
\value{
Resampled data
}
\description{
Get data back to 10 Hz for accumulation
}
\references{
Ali Neishabouri et al. DOI: https://doi.org/10.21203/rs.3.rs-1370418/v1
}
\author{
Jairo Hidalgo Migueles
}
