% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_30hz.R
\name{resample_30hz}
\alias{resample_30hz}
\title{resample_30hz}
\usage{
resample_30hz(raw = c(), sf = 30, verbose = FALSE)
}
\arguments{
\item{raw}{Matrix containing raw data}

\item{sf}{Sample frequency of raw data (Hz)}

\item{verbose}{Print diagnostic messages}
}
\value{
resampled_data
}
\description{
Resample the raw data.
}
\references{
Ali Neishabouri et al. DOI: https://doi.org/10.21203/rs.3.rs-1370418/v1
}
\author{
Jairo Hidalgo Migueles
}
