% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_counts.R
\name{sum_counts}
\alias{sum_counts}
\title{sum_counts}
\usage{
sum_counts(downsample_10hz, epoch = 60, verbose = FALSE)
}
\arguments{
\item{downsample_10hz}{Matrix containing downsampled to 10hz data}

\item{epoch}{Used to compute how many raw samples are used for computing an epoch}

\item{verbose}{Used to compute how many raw samples are used for computing an epoch}
}
\value{
Matrix with counts per epoch in the 3 axes
}
\description{
Generate counts per epoch.
}
\references{
Ali Neishabouri et al. DOI: https://doi.org/10.21203/rs.3.rs-1370418/v1
}
\author{
Jairo Hidalgo Migueles
}
