% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_data.R
\name{trim_data}
\alias{trim_data}
\title{trim_data}
\usage{
trim_data(bpf_data = c(), lfe_select = FALSE, verbose = FALSE)
}
\arguments{
\item{bpf_data}{Matrix containing filtered data}

\item{lfe_select}{False for regular trimming, True for allow more noise}

\item{verbose}{Print diagnostic messages}
}
\value{
The trimmed/thresholded data
}
\description{
trim_data
}
\references{
Ali Neishabouri et al. DOI: https://doi.org/10.21203/rs.3.rs-1370418/v1
}
\author{
Jairo Hidalgo Migueles
}
